+++ crates/imap-proto/src/protocol/enable.rs
@@ -21,10 +21,32 @@
  * for more details.
 */
 
-use super::capability::Capability;
+use super::{capability::Capability, ImapResponse};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Arguments {
     pub tag: String,
     pub capabilities: Vec<Capability>,
+}
+
+pub struct Response {
+    pub enabled: Vec<Capability>,
+}
+
+impl ImapResponse for Response {
+    fn serialize(self) -> Vec<u8> {
+        if !self.enabled.is_empty() {
+            let mut buf = Vec::with_capacity(64);
+            buf.extend(b"* ENABLED");
+            for capability in self.enabled {
+                buf.push(b' ');
+                capability.serialize(&mut buf);
+            }
+            buf.push(b'\r');
+            buf.push(b'\n');
+            buf
+        } else {
+            Vec::new()
+        }
+    }
 }